requirements.txt
Flask==2.0.3
numpy==1.21.2
pandas==1.4.1
plotly==5.6.0
python-dotenv==0.20.0
